Maximum compatibility and RSA 4096-bit keys required
High level changes
------------------

Better compatibility with more openssl versions.  The salt has been
lowered to 8-bytes.  The AES arguments remove -saltlen.  OpenSSL 3.2 is
no longer a hard requirement.

The password length has been increased to the maximum possible password
within the limits of 4096-bit RSA keys.  389 characters.

passin, salt, and hash have been combined into one encrypted value under
hash in the cipher YAML.

A human friendly error message is displayed if the password length is too
long.

New environment variable: pbkdf2_password_length so that 2048-bit RSA is
an option for those who need it.

Renamed environment variable openssl_saltlen to pbkdf2_salt_length for
clarity.

Additional security checks are performed when decrypting to ensure
2048-bit RSA keys are the minimum.

Deeper Technical explanation
----------------------------

hash and salt is stored as binary.  This improves security by allowing
larger passwords or salts in the new "combined hash".  The file size of
the cipher YAML is reduced as well.

RSA has limits to the amount of data that can be encrypted.  The limit is
defined as follows:

<key size in bits>/8-66 = <data limit in bytes>

- 4096-bit keys can encrypt 446 bytes
- 2048-bit keys can encrypt 190 bytes

Reserve data for hash and salt:

- Max salt size is 32 bytes ascii.
- Hash size is 64 bytes ascii.
- Hash and salt are converted to binary and rendered as 56 bytes of
  base64 chars.
- There is one ascii separator (1 byte)

Because of this, The reserve size for the hash field is 57 bytes (56+1).

So the password limit calculation is:

<key size in bits>/8-66-57

- 4096-bit password limit is 389 chars (4096/8-123)
- 2048-bit password limit is 133 chars (2048/8-123)